# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marcenaria API",
        "description": "Documentação da API da Marcenaria usando Laravel + Sanctum + Swagger",
        "contact": {
            "email": "suporte@marcenaria.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista todos as postagens",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Lista de postagens"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cria uma nova postagem",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "slug"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Projeto Teste"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "projeto-teste"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Conteúdo da postagem"
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "example": "Url da imagem"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "Tags da postagem"
                                    },
                                    "is_published": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postagem criada com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe uma postagem específica",
                "operationId": "142b5adb0520cca80cedfe09d9298332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da postagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postagem encontrada"
                    },
                    "404": {
                        "description": "Postagem não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualiza uma postagem",
                "operationId": "650f4bed0846d9350903d8a17c7b503a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da postagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Postagem Atualizada"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Novo conteúdo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Postagem atualizada com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Postagem não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exclui uma postagem",
                "operationId": "9d99f2e42af2f722a058a86245aa856f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da postagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Postagem excluída com sucesso"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Postagem não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use o token gerado no login como Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}